Add 'force' to crypttab
For some reason, the s390x system may fail to create the by-uuid links
for the disks, and then dracut may ignore those crypttab entries with
UUID when generating the initrd with '--hostonly'.

To ensure the LUKS partitions to be included, the 'force' option is
added so that dracut will ignore the absence of the by-uuid links.

Signed-off-by: Gary Lin <glin@suse.com>